+++ src/enums/route-names.enum.ts
@@ -9,6 +9,4 @@ export enum ROUTE_NAMES {
   nftItemEdit = 'nft-item-edit',
   nftsCreate = 'nfts-create',
   nftsOverview = 'nfts-overview',
-  kycRequests = 'kyc-requests',
-  settings = 'settings',
 }