+++ src/fields/FileField.vue
@@ -1,10 +1,9 @@
 <script lang="ts" setup>
 import { useDropZone } from '@vueuse/core'
 import { computed, ref } from 'vue'
-import { useFile } from '@/composables'
+import { useFile, useContext } from '@/composables'
 import { Icon, AppButton } from '@/common'
 import { Bus } from '@/helpers'
-import { useI18n } from 'vue-i18n'
 import { ErrorHandler } from '@/helpers/error-handler'
 import { formatBytes } from '@/helpers'
 import { Document } from '@/api'
@@ -35,7 +34,7 @@ const emit = defineEmits<{
   (e: 'update:modelValue', value: Document): void
 }>()
 
-const { t } = useI18n()
+const { $t } = useContext()
 
 const dropZoneRef = ref<HTMLDivElement | null>(null)
 
@@ -97,7 +96,7 @@ const onChange = (file: File) => {
       .join(', ')
 
     Bus.error(
-      t('file-field.incorrect-file-type-err', {
+      $t('file-field.incorrect-file-type-err', {
         allowedTypes: acceptedExtensions,
         type: `.${getFileExtension(file)}`,
       }),
@@ -107,7 +106,7 @@ const onChange = (file: File) => {
 
   if (file.size > maxSizeBytes.value) {
     Bus.error(
-      t('file-field.max-size-exceeded-err', {
+      $t('file-field.max-size-exceeded-err', {
         maxSize: props.maxSize,
       }),
     )