+++ src/main.ts
@@ -6,29 +6,20 @@ import log from 'loglevel'
 
 import VueToastificationPlugin from 'vue-toastification'
 import { ICON_NAMES, ROUTE_NAMES } from '@/enums'
-import { createApp, getCurrentInstance, h } from 'vue'
-import { useI18n } from 'vue-i18n'
+import { createApp } from 'vue'
 import { i18n } from '@/localization'
 import { config } from '@config'
 import { router } from '@/router'
 import { store } from '@/store'
 import MountedTeleport from '@/common/MountedTeleport.vue'
 
-const app = createApp({
-  setup() {
-    const app = getCurrentInstance()
-    const { t, locale } = useI18n({ useScope: 'global' })
-    if (app) {
-      app.appContext.config.globalProperties.$t = t
-      app.appContext.config.globalProperties.$locale = locale
-    }
-  },
-  render: () => h(App),
-})
-
 log.setDefaultLevel(config.LOG_LEVEL)
 
-app.use(router).use(store).use(i18n).use(VueToastificationPlugin)
+const app = createApp(App)
+  .use(router)
+  .use(store)
+  .use(i18n)
+  .use(VueToastificationPlugin)
 
 app.config.globalProperties.$routes = ROUTE_NAMES
 app.config.globalProperties.$config = config